+++ rdmo/core/settings.py
@@ -165,7 +165,14 @@ STATICFILES_FINDERS = (
     'compressor.finders.CompressorFinder',
 )
 
-STATICFILES_STORAGE = 'rdmo.core.storage.VersionedStaticFilesStorage'
+STORAGES = {
+    "default": {
+        "BACKEND": "django.core.files.storage.FileSystemStorage",
+    },
+    "staticfiles": {
+        "BACKEND": "rdmo.core.storage.VersionedStaticFilesStorage",
+    },
+}
 
 DATABASES = {
     'default': {