+++ rdmo/questions/tests/test_queries_catalog.py
@@ -11,10 +11,10 @@ max_queries = [
     ('list', 8, {}, {}),
     ('index', 3, {}, {}),
     ('nested', 31, {'pk': 1}, {}),
-    ('export', 32, {'export_format': 'xml'}, {}),
+    ('export', 36, {'export_format': 'xml'}, {}),
     ('export', 42, {'export_format': 'xml'}, {'full': '1'}),
     ('detail', 8, {'pk': 1}, {}),
-    ('detail_export', 32, {'pk': 1, 'export_format': 'xml'}, {}),
+    ('detail_export', 36, {'pk': 1, 'export_format': 'xml'}, {}),
     ('detail_export', 39, {'pk': 1, 'export_format': 'xml'}, {'full': '1'}),
 ]