+++ rdmo/questions/tests/test_queries_question.py
@@ -10,10 +10,10 @@ max_queries = [
     # action, max_queries, url_kwargs, url_params
     ('list', 8, {}, {}),
     ('index', 3, {}, {}),
-    ('export', 10, {'export_format': 'xml'}, {}),
+    ('export', 12, {'export_format': 'xml'}, {}),
     ('export', 13, {'export_format': 'xml'}, {'full': '1'}),
     ('detail', 8, {'pk': 1}, {}),
-    ('detail_export', 6, {'pk': 1, 'export_format': 'xml'}, {}),
+    ('detail_export', 8, {'pk': 1, 'export_format': 'xml'}, {}),
     ('detail_export', 9, {'pk': 1, 'export_format': 'xml'}, {'full': '1'}),
 ]