+++ rdmo/questions/tests/test_queries_section.py
@@ -11,10 +11,10 @@ max_queries = [
     ('list', 7, {}, {}),
     ('index', 3, {}, {}),
     ('nested', 11, {'pk': 1}, {}),
-    ('export', 30, {'export_format': 'xml'}, {}),
+    ('export', 34, {'export_format': 'xml'}, {}),
     ('export', 40, {'export_format': 'xml'}, {'full': '1'}),
     ('detail', 7, {'pk': 1}, {}),
-    ('detail_export', 14, {'pk': 1, 'export_format': 'xml'}, {}),
+    ('detail_export', 16, {'pk': 1, 'export_format': 'xml'}, {}),
     ('detail_export', 17, {'pk': 1, 'export_format': 'xml'}, {'full': '1'}),
 ]